# historian/backend/src/main/swagger/api-swagger.yml
swagger: '2.0'
info:
  description: REST API for logisland historian
  version: v1
  title: Logisland Data Historian
  contact:
    name: hurence
    email: support@hurence.com
host: localhost:8701
basePath: /api/v1
schemes:
  - http
  - https
paths:
...
  /selections:
      get:
        tags:
          - selection
        operationId: getAllUserSelection
        summary:  get all slections that this user has access to
        description: retrieve all slections that user has access to
        produces:
          - application/json
        responses:
          '200':
             description: Selection list
             schema:
               type: array
               items:
                 $ref: '#/definitions/Selection'
          default:
             description: unexpected error
             schema:
               $ref: '#/definitions/Error'
  /selections/{selectionId}:
      get:
       tags:
         - selection
       summary: get Selection
       description: get the corresponding Selection
       operationId: getSelection
       produces:
         - application/json
       parameters:
         - name: selectionId
           in: path
           description: id of the Selection to return
           required: true
           type: string
       responses:
         '200':
           description: Selection found
           schema:
             $ref: '#/definitions/Selection'
         '404':
            description: Selection resource not found
         default:
           description: unexpected error
           schema:
             $ref: '#/definitions/Error'
      delete:
       tags:
         - selection
       summary: delete Selection
       description: remove the corresponding Selection
       operationId: deleteSelection
       parameters:
         - name: selectionId
           in: path
           description: id of the Selection to be deleted
           required: true
           type: string
       responses:
         '404':
           description: Selection resource not found
         '400':
           description: Invalid ID supplied
         '200':
           description: Selection successfully removed
           schema:
             $ref: '#/definitions/Selection'
         default:
           description: unexpected error
           schema:
             $ref: '#/definitions/Error'
      put:
       tags:
        - selection
       summary:  update Selection
       description: update an existing Selection
       operationId: updateSelection
       parameters:
         - name: selectionId
           in: path
           description: selectionId to be updated
           required: true
           type: string
         - name: body
           in: body
           description: new Selection definition
           required: true
           schema:
             $ref: '#/definitions/Selection'
       responses:
        '404':
          description: Selection resource not found
        '400':
          description: Invalid ID supplied
        '200':
         description: Selection successfuly updated
         schema:
           $ref: '#/definitions/Selection'
      post:
       tags:
        - selection
       summary:  create new Selection
       description:  store a new Selection
       operationId: addSelectionWithId
       parameters:
         - name: body
           in: body
           description: Selection resource to add
           required: true
           schema:
             $ref: '#/definitions/Selection'
         - name: selectionId
           in: path
           description: selectionId to create
           required: true
           type: string
       responses:
         '400':
           description: Invalid ID supplied
         '200':
           description: Selection successfuly created
           schema:
             $ref: '#/definitions/Selection'
         default:
           description: unexpected error
           schema:
             $ref: '#/definitions/Error'
  /selections/{selectionId}/tags:
       get:
         tags:
          - selection
          - tag
         operationId: getAllTagsFromSelection
         summary:  get all tags from a selection
         description: retrieve all tags through this selection
         produces:
            - application/json
         parameters:
           - name: selectionId
             in: path
             description: id of the selection
             required: true
             type: string
         responses:
           '200':
             description: Tag list
             schema:
               type: array
               items:
                 $ref: '#/definitions/Tag'
           default:
             description: unexpected error
             schema:
               $ref: '#/definitions/Error'
...
  /pipelines:
    get:
      tags:
       - pipeline
      operationId: getAllPipelines
      summary:  get all pipelines
      description: retrieve all pipelines (retrieve only summary fields)
      responses:
        "200":
          description: pipeline configuration list
          schema:
            type: array
            items:
              $ref: '#/definitions/Pipeline'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
  /pipelines/{pipelineId}:
    get:
      tags:
        - pipeline
      summary: get pipeline
      description: get the corresponding Pipeline definition
      operationId: getPipeline
      produces:
        - application/json
        - text/plain
      parameters:
        - name: pipelineId
          in: path
          description: id of the pipeline to return
          required: true
          type: string
      responses:
        "200":
          description: pipeline definition
          schema:
            $ref: '#/definitions/Pipeline'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
    delete:
      tags:
        - pipeline
      summary: delete pipeline
      description: remove the corresponding Pipeline definition and stop if its currently running
      operationId: deletePipeline
      parameters:
        - name: pipelineId
          in: path
          description: id of the pipeline to return
          required: true
          type: string
      responses:
        "404":
          description: Pipeline not found
        "400":
          description: Invalid ID supplied
        "200":
          description: pipeline successfully removed
          schema:
            $ref: '#/definitions/Pipeline'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
    put:
      tags:
       - pipeline
      summary:  update pipeline
      description: update an existing pipeline configuration if valid
      operationId: updatePipeline
      parameters:
        - name: pipelineId
          in: path
          description: Pipeline to add to the store
          required: true
          type: string
        - name: pipeline
          in: body
          description: Pipeline to add to the store
          required: true
          schema:
            $ref: '#/definitions/Pipeline'
      responses:
        "200":
          description: Pipeline successfuly created
          schema:
            $ref: '#/definitions/Pipeline'
  /pipelines/{pipelineId}/start:
     post:
      tags:
        - pipeline
      summary: start pipeline
      description: start the corresponding Pipeline definition
      operationId: startPipeline
      parameters:
        - name: pipelineId
          in: path
          description: id of the pipeline to return
          required: true
          type: string
      responses:
        "200":
          description: pipeline successfuly started
          schema:
            $ref: '#/definitions/Pipeline'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
  /pipelines/{pipelineId}/stop:
    post:
      tags:
       - pipeline
      summary:  create new pipeline
      description:  store a new pipeline configuration if valid
      operationId: addPipelineWithId
      parameters:
        - name: body
          in: body
          description: Pipeline configuration to add to the store
          required: true
          schema:
            $ref: '#/definitions/Pipeline'
        - name: pipelineId
          in: path
          description: PipelineId to add to the store
          required: true
          type: string
      responses:
        "404":
          description: Pipeline not found
        "400":
          description: Invalid ID supplied
        "200":
          description: Pipeline successfuly created
          schema:
            $ref: '#/definitions/Pipeline'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
definitions:
...
  Selection:
    type: object
    description: a Selection is a selection of tags.
    required:
      - id
      - tags
    properties:
      id:
        description: concatenation of owner with name for example
        type: string
      name:
        type: string
      owner:
        type: string
      tags:
        description: list of tags id in selection
        type: array
        items:
          type: string
      record_type:
        type: string
        default: 'selection'
  Tag:
    type: object
    description: a Tag is an identifier to an OPC value
    required:
      - item_name
      - datasource_id
      - data_type
    properties:
      record_type:
        type: string
        default: 'tag'
      id:
        type: string
        default: 'mySweetUniqueId'
      datasource_id:
        type: string
      domain:
        type: string
      server:
        type: string
      group:
        type: string
      tag_name:
        type: string
      labels:
        description: can be used to group tags around a theme
        type: array
        items:
          type: string
      data_type:
        type: string
        default: float
        enum:
        - int
        - long
        - float
        - double
        - string
        - array
        - null
        - bytes
        - boolean
      description:
        type: string
      text:
        description: ctach all field
        type: array
        items:
          type: string
      creation_date:
        description: timestamp
        type: integer
        format: int64
      last_modification_date:
        description: timestamp
        type: integer
        format: int64
      last_polling_date:
        description: timestamp
        type: integer
        format: int64
      update_rate:
        description: polling delay in ms
        default: 1000
        type: integer
      min_numeric_value:
        type: number
        format: dumble
      max_numeric_value:
        type: number
        format: double
      last_numeric_value:
        type: number
        format: double
      last_quality:
        type: integer
...
  Pipeline:
    type: object
    description: a Pipeline is a sequence of Processor wich apply specifi business logic on current tag value (matching query for alerting for example)
    required:
      - component
      - name
    properties:
      record_type:
        type: string
        default: 'pipeline'
      name:
        type: string
      component:
        type: string
      documentation:
        type: string
      config:
        type: array
        items:
          $ref: '#/definitions/Property'
      processors:
        type: array
        items:
          $ref: '#/definitions/Processor'
  Processor:
    type: object
    required:
      - component
      - config
      - name
    properties:
      name:
        type: string
      component:
        type: string
      documentation:
        type: string
      config:
        type: array
        items:
          $ref: '#/definitions/Property'
  Property:
    type: object
    required:
      - key
      - value
    properties:
      key:
        type: string
      type:
        type: string
        default: "string"
      value:
        type: string
  Error:
    required:
      - code
      - message
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string